feat: Add renderCustomStats prop to render extra stats & move storage and version to renderCustomStats
